+++ flutter_inappwebview/example/integration_test/proxy_controller/main.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
@@ -8,11 +9,13 @@ import '../env.dart';
 import '../util.dart';
 
 part 'clear_and_set_proxy_override.dart';
+part 'malformed_rules.dart';
 
 void main() {
   final shouldSkip = !ProxyController.isClassSupported();
 
   skippableGroup('Proxy Controller', () {
     clearAndSetProxyOverride();
+    malformedRules();
   }, skip: shouldSkip);
 }